Renamed Market sub-directories
